refactor: 자료형 변환(BE)

+++ clean4u/src/main/java/org/example/clean4u/order/OrderRepository.java
@@ -9,7 +9,7 @@ public interface OrderRepository extends OrderRepositoryCustom, JpaRepository<Or
 
     List<Order> findByCustomerIdOrderByOrderDateDesc(Long customer);
 
-    Long countByStatus(OrderStatus status);
+    Integer countByStatus(OrderStatus status);
 
     List<Order> findByStatus(OrderStatus status);
 }
